chore: move options in middleware file
